requirements.txt
anyio==3.6.2
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
cssselect==1.2.0
Cython==0.29.34
fake-useragent==1.1.3
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
lxml==4.9.2
numpy==1.24.3
pandas==2.0.1
parse==1.19.0
psycopg2-binary
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
requests-html==0.10.0
selectolax==0.3.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
